Cast indentation counter to int to fix compilation on Mac OS X
The error was:

sgml-parser.c: In function 'print_indent':
sgml-parser.c:99: warning: field precision should have type 'int', but argument 2 has type 'long unsigned int'

+++ src/dom/test/sgml-parser.c
@@ -96,7 +96,7 @@ static unsigned char indent_string[] =
 static void
 print_indent(struct dom_stack *stack)
 {
-	printf("%.*s", get_indent_offset(stack), indent_string);
+	printf("%.*s", (int) get_indent_offset(stack), indent_string);
 }
 
 static enum dom_code
